sdm845-common: Include sphal library for graphics mapper 4.0
Nothing uses this yet but the new HAL provides allocator@4.0 when
TARGET_USES_GRALLOC4 is not explicitly set to false. Complement this by
also including mapper@4.0, in case we get blobs that use it in the
future.

(Currently both Adreno blobs and the opensource composer use 3.0 with a
fallback to 2.0)

Signed-off-by: Marijn Suijten <marijns95@gmail.com>
Change-Id: I4f277cdc6f57a2b33039e30cf2142aeda07c66f8